+++ backend/controllers/CompanyController.php
@@ -8,6 +8,8 @@ use app\models\CompanySearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\web\UploadedFile;
+use yii\web\ForbiddenHttpException;
 
 /**
  * CompanyController implements the CRUD actions for Company model.
@@ -35,6 +37,7 @@ class CompanyController extends Controller
      */
     public function actionIndex()
     {
+       // echo '<pre>';print_r(Yii::$app->user);exit();
         $searchModel = new CompanySearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -63,14 +66,30 @@ class CompanyController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Company();
+      // echo '<pre>';print_r(Yii::$app->user->identity);exit();
+        if (Yii::$app->user->can( 'create-company' )) {
+           $model = new Company();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->company_id]);
-        } else {
-            return $this->render('create', [
-                'model' => $model,
-            ]);
+            if ($model->load(Yii::$app->request->post())) {
+                $file=$_FILES;
+               // echo '<pre>';print_r($file['Company']['name']['file']);exit();
+              //  echo $model->$file->extension;exit();
+                $ImageName =$file['Company']['name']['file'];
+                $model->file = UploadedFile::getInstance($model,'file');
+               // echo '<pre>';print_r($model->file);exit();
+                $model->file->saveAs( 'uploads/'.$ImageName );
+
+                $model->logo = 'uploads/'.$ImageName;
+                $model->save();
+
+                return $this->redirect(['view', 'id' => $model->company_id]);
+            } else {
+                return $this->render('create', [
+                    'model' => $model,
+                ]);
+            }
+        }else{
+            throw new ForbiddenHttpException("You don't have permission to access this page.");
         }
     }
 
@@ -82,14 +101,25 @@ class CompanyController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if (Yii::$app->user->can( 'edit-company' )) {
+            $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->company_id]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+
+                $ImageName = $model->company_name;
+                $model->file = UploadedFile::getInstance($model, 'file');
+                $model->file->saveAs( 'uploads/'.$ImageName.'.'.$model->$file->extension );
+
+               //echo $model->logo = 'uploads/'.$ImageName.'.'.$model->$file->extension;
+    //die;
+                return $this->redirect(['view', 'id' => $model->company_id]);
+            } else {
+                return $this->render('update', [
+                    'model' => $model,
+                ]);
+            }
+         }else{
+            throw new ForbiddenHttpException("You don't have permission to access this page.");
         }
     }
 
@@ -101,9 +131,14 @@ class CompanyController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if (Yii::$app->user->can( 'delete-company' )) {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+         }else{
+            throw new ForbiddenHttpException("You don't have permission to access this page.");
+        }
+
     }
 
     /**
